Support 16K sequence coding
Fixed a couple of variable/function definitions to support 16K
sequence coding at high bit-rates.

+++ vp9/encoder/vp9_onyx_if.c
@@ -1090,13 +1090,13 @@ void vp9_new_frame_rate(VP9_COMP *cpi, double framerate) {
 }
 
 
-static int
-rescale(int val, int num, int denom) {
+static int64_t
+rescale(int val, int64_t num, int denom) {
   int64_t llnum = num;
   int64_t llden = denom;
   int64_t llval = val;
 
-  return (int)(llval * llnum / llden);
+  return (int64_t)(llval * llnum / llden);
 }
 
 static void set_tile_limits(VP9_COMP *cpi) {
